+++ momoim-core/src/main/java/com/triplem/momoim/core/domain/review/ReviewRepositoryImpl.java
@@ -8,6 +8,8 @@ import static com.triplem.momoim.core.domain.user.QUserEntity.userEntity;
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.triplem.momoim.core.common.PaginationInformation;
+import com.triplem.momoim.core.domain.review.dto.MyReview;
+import com.triplem.momoim.core.domain.review.dto.ReviewDetail;
 import com.triplem.momoim.exception.BusinessException;
 import com.triplem.momoim.exception.ExceptionCode;
 import java.util.List;
