+++ momoim-external-api/src/main/java/com/triplem/momoim/api/review/controller/ReviewController.java
@@ -8,9 +8,9 @@ import com.triplem.momoim.api.review.service.ReviewService;
 import com.triplem.momoim.auth.utils.SecurityUtil;
 import com.triplem.momoim.core.common.PaginationInformation;
 import com.triplem.momoim.core.domain.gathering.dto.GatheringPreview;
-import com.triplem.momoim.core.domain.review.MyReview;
 import com.triplem.momoim.core.domain.review.Review;
-import com.triplem.momoim.core.domain.review.ReviewDetail;
+import com.triplem.momoim.core.domain.review.dto.MyReview;
+import com.triplem.momoim.core.domain.review.dto.ReviewDetail;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
