+++ momoim-external-api/src/main/java/com/triplem/momoim/api/review/service/ReviewService.java
@@ -3,12 +3,12 @@ package com.triplem.momoim.api.review.service;
 import com.triplem.momoim.core.common.PaginationInformation;
 import com.triplem.momoim.core.domain.gathering.dto.GatheringPreview;
 import com.triplem.momoim.core.domain.gathering.infrastructure.GatheringRepository;
-import com.triplem.momoim.core.domain.review.ModifyReview;
-import com.triplem.momoim.core.domain.review.MyReview;
 import com.triplem.momoim.core.domain.review.Review;
-import com.triplem.momoim.core.domain.review.ReviewDetail;
 import com.triplem.momoim.core.domain.review.ReviewRegister;
 import com.triplem.momoim.core.domain.review.ReviewRepository;
+import com.triplem.momoim.core.domain.review.dto.ModifyReview;
+import com.triplem.momoim.core.domain.review.dto.MyReview;
+import com.triplem.momoim.core.domain.review.dto.ReviewDetail;
 import com.triplem.momoim.exception.BusinessException;
 import com.triplem.momoim.exception.ExceptionCode;
 import java.util.List;
